requirements_test.txt
pytest==8.3.5
pytest-homeassistant-custom-component==0.13.233
